refact: auto reconnect logic
