+++ hooks/room.ts
@@ -11,8 +11,6 @@ import {
 import { DebugContext } from 'contexts/DebugContext';
 import { TelnyxMeetContext } from 'contexts/TelnyxMeetContext';
 
-const TOKEN_TTL = 50;
-
 interface Props {
   roomId: string;
   tokens: {
@@ -22,7 +20,7 @@ interface Props {
   context: {};
   callbacks?: {
     onConnected?: () => void;
-    onDisconnected?: () => void;
+    onDisconnected?: (reason: string) => void;
     onParticipantJoined?: (
       participantId: Participant['id'],
       state: State
@@ -95,168 +93,151 @@ export const useRoom = ({
     _setMessages(messagesRef.current);
   };
 
-  useEffect(() => {
-    const connectAndJoinRoom = async () => {
-      try {
-        roomRef.current = await initialize({
-          roomId,
-          clientToken,
-          context: JSON.stringify(context),
-          logLevel: 'DEBUG',
-          enableMessages: true,
-        });
-      } catch (error) {
-        sendNotification({ body: (error as Error).message });
-        typeof callbacks?.onDisconnected === 'function' &&
-          callbacks.onDisconnected();
-        roomRef.current = undefined;
-      }
-
-      // failed to initialize don't proceed further
-      if (!roomRef.current) {
-        return;
-      }
-
-      setState(roomRef.current.getState());
-      setDebugState(roomRef.current.getState());
-
-      roomRef.current.on('state_changed', (value) => {
-        setState(value);
-        setDebugState(value);
+  const connectAndJoinRoom = async () => {
+    try {
+      roomRef.current = await initialize({
+        roomId,
+        clientToken,
+        context: JSON.stringify(context),
+        logLevel: 'DEBUG',
+        enableMessages: true,
       });
+    } catch (error) {
+      sendNotification({ body: (error as Error).message });
+      typeof callbacks?.onDisconnected === 'function' &&
+        callbacks.onDisconnected('network_error');
 
-      roomRef.current.on('connected', (state) => {
-        setParticipantsByActivity((value) => {
-          return new Set([
-            roomRef.current!.getLocalParticipant().id,
-            ...state.participants.keys(),
-          ]);
-        });
-        state.streams.forEach((stream) => {
-          if (stream.key === 'presentation') {
-            setPresenter(state.participants.get(stream.participantId));
-          }
+      // setTimeout(async () => {
+      //   sendNotification({
+      //     body: 'catch==>Auto-reconnect - Trying to reconnect',
+      //   });
 
-          if (
-            stream.participantId === roomRef.current!.getLocalParticipant().id
-          ) {
-            return;
-          }
+      //   await connectAndJoinRoom();
+      // }, 3000);
+    }
 
-          roomRef.current!.addSubscription(stream.participantId, stream.key, {
-            audio: true,
-            video: true,
-          });
-        });
-        typeof callbacks?.onConnected === 'function' && callbacks.onConnected();
+    // failed to initialize don't proceed further
+    if (!roomRef.current) {
+      return;
+    }
+
+    setState(roomRef.current.getState());
+    setDebugState(roomRef.current.getState());
+
+    roomRef.current.on('state_changed', (value) => {
+      setState(value);
+      setDebugState(value);
+    });
+
+    roomRef.current.on('connected', (state) => {
+      setParticipantsByActivity((value) => {
+        return new Set([
+          roomRef.current!.getLocalParticipant().id,
+          ...state.participants.keys(),
+        ]);
+      });
+      state.streams.forEach((stream) => {
+        if (stream.key === 'presentation') {
+          setPresenter(state.participants.get(stream.participantId));
+        }
 
         if (
-          optionalFeatures.isNetworkMetricsEnabled &&
-          state.participants.size > 0 &&
-          roomRef.current!.getLocalParticipant().id
+          stream.participantId === roomRef.current!.getLocalParticipant().id
         ) {
-          const participantIds: Array<string> = [];
-
-          state.participants.forEach((item) => {
-            participantIds.push(item.id);
-          });
-
-          roomRef.current!.enableNetworkMetricsReport(participantIds);
+          return;
         }
+
+        roomRef.current!.addSubscription(stream.participantId, stream.key, {
+          audio: true,
+          video: true,
+        });
       });
+      typeof callbacks?.onConnected === 'function' && callbacks.onConnected();
 
-      roomRef.current.on('disconnected', (reason, state) => {
-        if (reason === 'network_error') {
-          sendNotification({
-            body: 'network_error - Check your network connection',
-          });
-        }
+      if (
+        optionalFeatures.isNetworkMetricsEnabled &&
+        state.participants.size > 0 &&
+        roomRef.current!.getLocalParticipant().id
+      ) {
+        const participantIds: Array<string> = [];
 
-        if (reason === 'user_initiated') {
-          sendNotification({
-            body: 'user_initiated - Disconnect method executed by user',
-          });
-        }
-
-        if (reason === 'timed_out') {
-          sendNotification({
-            body: 'timed_out - Reached out timeout limit to connect',
-          });
-        }
-
-        setParticipantsByActivity(new Set());
-        typeof callbacks?.onDisconnected === 'function' &&
-          callbacks.onDisconnected();
-      });
-
-      roomRef.current.on('participant_joined', (participantId, state) => {
-        if (
-          optionalFeatures.isNetworkMetricsEnabled &&
-          state.participants.size > 0
-        ) {
-          const participantIds: Array<string> = [];
-          state.participants.forEach((item) => {
-            participantIds.push(item.id);
-          });
-
-          roomRef.current!.enableNetworkMetricsReport(participantIds);
-        }
-
-        setParticipantsByActivity((value) => {
-          return new Set([
-            roomRef.current!.getLocalParticipant().id,
-            ...value,
-            participantId,
-          ]);
+        state.participants.forEach((item) => {
+          participantIds.push(item.id);
         });
 
-        typeof callbacks?.onParticipantJoined === 'function' &&
-          callbacks.onParticipantJoined(participantId, state);
+        roomRef.current!.enableNetworkMetricsReport(participantIds);
+      }
+    });
+
+    roomRef.current.on('disconnected', (reason, state) => {
+      if (reason === 'network_error') {
+        sendNotification({
+          body: 'network_error - Check your network connection',
+        });
+      }
+
+      if (reason === 'user_initiated') {
+        sendNotification({
+          body: 'user_initiated - Disconnect method executed by user',
+        });
+      }
+
+      if (reason === 'timed_out') {
+        sendNotification({
+          body: 'timed_out - Reached out timeout limit to connect',
+        });
+      }
+
+      setParticipantsByActivity(new Set());
+      typeof callbacks?.onDisconnected === 'function' &&
+        callbacks.onDisconnected(reason);
+    });
+
+    roomRef.current.on('participant_joined', (participantId, state) => {
+      if (
+        optionalFeatures.isNetworkMetricsEnabled &&
+        state.participants.size > 0
+      ) {
+        const participantIds: Array<string> = [];
+        state.participants.forEach((item) => {
+          participantIds.push(item.id);
+        });
+
+        roomRef.current!.enableNetworkMetricsReport(participantIds);
+      }
+
+      setParticipantsByActivity((value) => {
+        return new Set([
+          roomRef.current!.getLocalParticipant().id,
+          ...value,
+          participantId,
+        ]);
       });
 
-      roomRef.current.on(
-        'participant_leaving',
-        (participantId, reason, state) => {
-          if (reason === 'kicked') {
-            if (state.localParticipantId === participantId) {
-              sendNotification({
-                body: 'You got kicked from the room by the moderator!',
-              });
-            } else {
-              const context = JSON.parse(
-                state.participants.get(participantId).context
-              );
+      typeof callbacks?.onParticipantJoined === 'function' &&
+        callbacks.onParticipantJoined(participantId, state);
+    });
 
-              sendNotification({
-                body: `${
-                  context.username ? context.username : participantId
-                } has been kicked by the moderator!`,
-              });
-            }
-          }
+    roomRef.current.on(
+      'participant_leaving',
+      (participantId, reason, state) => {
+        if (reason === 'kicked') {
+          if (state.localParticipantId === participantId) {
+            sendNotification({
+              body: 'You got kicked from the room by the moderator!',
+            });
+          } else {
+            const context = JSON.parse(
+              state.participants.get(participantId).context
+            );
 
-          if (
-            optionalFeatures.isNetworkMetricsEnabled &&
-            state.participants.size > 0
-          ) {
-            roomRef.current!.disableNetworkMetricsReport([participantId]);
+            sendNotification({
+              body: `${
+                context.username ? context.username : participantId
+              } has been kicked by the moderator!`,
+            });
           }
         }
-      );
-
-      roomRef.current.on('participant_left', (participantId, state) => {
-        if (presenter?.id === participantId) {
-          setPresenter(undefined);
-        }
-
-        if (dominantSpeakerId === participantId) {
-          setDominantSpeakerId(undefined);
-        }
-
-        setParticipantsByActivity((value) => {
-          value.delete(participantId);
-          return new Set([roomRef.current!.getLocalParticipant().id, ...value]);
-        });
 
         if (
           optionalFeatures.isNetworkMetricsEnabled &&
@@ -264,162 +245,177 @@ export const useRoom = ({
         ) {
           roomRef.current!.disableNetworkMetricsReport([participantId]);
         }
-      });
-
-      roomRef.current.on('stream_published', (participantId, key, state) => {
-        if (key === 'presentation') {
-          setPresenter(state.participants.get(participantId));
-        }
-
-        if (participantId === roomRef.current!.getLocalParticipant().id) {
-          return;
-        }
-
-        roomRef.current!.addSubscription(participantId, key, {
-          audio: true,
-          video: true,
-        });
-      });
-
-      roomRef.current.on('stream_unpublished', (participantId, key, state) => {
-        if (key === 'presentation') {
-          setPresenter(undefined);
-        }
-
-        if (dominantSpeakerId === participantId && key === 'self') {
-          setDominantSpeakerId(undefined);
-        }
-      });
-
-      roomRef.current.on('track_enabled', (participantId, key, kind, state) => {
-        if (kind === 'video') {
-          setIsVideoPlaying(true);
-        }
-      });
-      roomRef.current.on(
-        'track_disabled',
-        (participantId, key, kind, state) => {
-          if (kind === 'video') {
-            setIsVideoPlaying(false);
-          }
-        }
-      );
-
-      roomRef.current.on(
-        'track_censored',
-        (participantId, key, kind, state) => {
-          if (kind === 'audio') {
-            if (state.localParticipantId === participantId) {
-              sendNotification({
-                body: `Your audio from "${key}" stream has been censored by the moderator`,
-              });
-            } else {
-              const context = JSON.parse(
-                state.participants.get(participantId).context
-              );
-
-              sendNotification({
-                body: `${
-                  context.username ? context.username : participantId
-                }'s audio from "${key}" stream has been censored by the moderator`,
-              });
-            }
-          }
-        }
-      );
-
-      roomRef.current.on(
-        'track_uncensored',
-        (participantId, key, kind, state) => {
-          if (kind === 'audio') {
-            if (state.localParticipantId === participantId) {
-              sendNotification({
-                body: `Your audio from "${key}" stream has been uncensored by the moderator`,
-              });
-            } else {
-              const context = JSON.parse(
-                state.participants.get(participantId).context
-              );
-
-              sendNotification({
-                body: `${
-                  context.username ? context.username : participantId
-                }'s audio from "${key}" stream has been uncensored by the moderator`,
-              });
-            }
-          }
-        }
-      );
-
-      roomRef.current.on('audio_activity', (participantId, key) => {
-        if (
-          (!key || key === 'self') &&
-          participantId !== roomRef.current!.getLocalParticipant().id
-        ) {
-          console.log(`${participantId} is speaking`);
-          setDominantSpeakerId(participantId);
-          setParticipantsByActivity((value) => {
-            return new Set([
-              roomRef.current!.getLocalParticipant().id,
-              participantId,
-              ...value,
-            ]);
-          });
-        }
-      });
-
-      roomRef.current.on(
-        'subscription_started',
-        (participantId, key, state) => {}
-      );
-      roomRef.current.on(
-        'subscription_reconfigured',
-        (participantId, key, state) => {}
-      );
-      roomRef.current.on(
-        'subscription_ended',
-        (participantId, key, state) => {}
-      );
-      roomRef.current.on(
-        'message_received',
-        (participantId, message, recipients, state) => {
-          const participant = state.participants.get(participantId);
-          const fromUsername = JSON.parse(participant.context).username;
-
-          if (!unreadMessages.current) {
-            unreadMessages.current = [];
-          }
-
-          const newMessages = unreadMessages.current.concat({
-            from: participantId,
-            fromUsername,
-            message,
-            recipients,
-          });
-
-          unreadMessages.current = newMessages;
-
-          setMessages({
-            from: participantId,
-            fromUsername,
-            message,
-            recipients,
-          });
-        }
-      );
-
-      roomRef.current.on('network_metrics_report', (networkMetrics) => {
-        console.debug('network_metrics_report', networkMetrics);
-        setNetworkMetrics(networkMetrics);
-      });
-
-      try {
-        await roomRef.current.connect();
-      } catch (error) {
-        throw error;
       }
-    };
+    );
 
+    roomRef.current.on('participant_left', (participantId, state) => {
+      if (presenter?.id === participantId) {
+        setPresenter(undefined);
+      }
+
+      if (dominantSpeakerId === participantId) {
+        setDominantSpeakerId(undefined);
+      }
+
+      setParticipantsByActivity((value) => {
+        value.delete(participantId);
+        return new Set([roomRef.current!.getLocalParticipant().id, ...value]);
+      });
+
+      if (
+        optionalFeatures.isNetworkMetricsEnabled &&
+        state.participants.size > 0
+      ) {
+        roomRef.current!.disableNetworkMetricsReport([participantId]);
+      }
+    });
+
+    roomRef.current.on('stream_published', (participantId, key, state) => {
+      if (key === 'presentation') {
+        setPresenter(state.participants.get(participantId));
+      }
+
+      if (participantId === roomRef.current!.getLocalParticipant().id) {
+        return;
+      }
+
+      roomRef.current!.addSubscription(participantId, key, {
+        audio: true,
+        video: true,
+      });
+    });
+
+    roomRef.current.on('stream_unpublished', (participantId, key, state) => {
+      if (key === 'presentation') {
+        setPresenter(undefined);
+      }
+
+      if (dominantSpeakerId === participantId && key === 'self') {
+        setDominantSpeakerId(undefined);
+      }
+    });
+
+    roomRef.current.on('track_enabled', (participantId, key, kind, state) => {
+      if (kind === 'video') {
+        setIsVideoPlaying(true);
+      }
+    });
+    roomRef.current.on('track_disabled', (participantId, key, kind, state) => {
+      if (kind === 'video') {
+        setIsVideoPlaying(false);
+      }
+    });
+
+    roomRef.current.on('track_censored', (participantId, key, kind, state) => {
+      if (kind === 'audio') {
+        if (state.localParticipantId === participantId) {
+          sendNotification({
+            body: `Your audio from "${key}" stream has been censored by the moderator`,
+          });
+        } else {
+          const context = JSON.parse(
+            state.participants.get(participantId).context
+          );
+
+          sendNotification({
+            body: `${
+              context.username ? context.username : participantId
+            }'s audio from "${key}" stream has been censored by the moderator`,
+          });
+        }
+      }
+    });
+
+    roomRef.current.on(
+      'track_uncensored',
+      (participantId, key, kind, state) => {
+        if (kind === 'audio') {
+          if (state.localParticipantId === participantId) {
+            sendNotification({
+              body: `Your audio from "${key}" stream has been uncensored by the moderator`,
+            });
+          } else {
+            const context = JSON.parse(
+              state.participants.get(participantId).context
+            );
+
+            sendNotification({
+              body: `${
+                context.username ? context.username : participantId
+              }'s audio from "${key}" stream has been uncensored by the moderator`,
+            });
+          }
+        }
+      }
+    );
+
+    roomRef.current.on('audio_activity', (participantId, key) => {
+      if (
+        (!key || key === 'self') &&
+        participantId !== roomRef.current!.getLocalParticipant().id
+      ) {
+        console.log(`${participantId} is speaking`);
+        setDominantSpeakerId(participantId);
+        setParticipantsByActivity((value) => {
+          return new Set([
+            roomRef.current!.getLocalParticipant().id,
+            participantId,
+            ...value,
+          ]);
+        });
+      }
+    });
+
+    roomRef.current.on(
+      'subscription_started',
+      (participantId, key, state) => {}
+    );
+    roomRef.current.on(
+      'subscription_reconfigured',
+      (participantId, key, state) => {}
+    );
+    roomRef.current.on('subscription_ended', (participantId, key, state) => {});
+    roomRef.current.on(
+      'message_received',
+      (participantId, message, recipients, state) => {
+        const participant = state.participants.get(participantId);
+        const fromUsername = JSON.parse(participant.context).username;
+
+        if (!unreadMessages.current) {
+          unreadMessages.current = [];
+        }
+
+        const newMessages = unreadMessages.current.concat({
+          from: participantId,
+          fromUsername,
+          message,
+          recipients,
+        });
+
+        unreadMessages.current = newMessages;
+
+        setMessages({
+          from: participantId,
+          fromUsername,
+          message,
+          recipients,
+        });
+      }
+    );
+
+    roomRef.current.on('network_metrics_report', (networkMetrics) => {
+      console.debug('network_metrics_report', networkMetrics);
+      setNetworkMetrics(networkMetrics);
+    });
+
+    try {
+      await roomRef.current.connect();
+    } catch (error) {
+      throw error;
+    }
+  };
+
+  useEffect(() => {
     if (!roomRef.current) {
       connectAndJoinRoom();
     }
@@ -470,7 +466,7 @@ export const useRoom = ({
         const data = await response.json();
         setClientToken(data.token);
       }
-    }, (TOKEN_TTL - 20) * 1000);
+    }, (50 - 20) * 1000);
 
     return () => {
       clearInterval(refreshTokenIntervalId);
